Create Data Service Class
Created a basic data service class that acts as convenient library of
functions that can be used to get data from the server. Currently just
using a stub so that I can add a basic unit test in a bit, but it seems
to be reasonable for the moment, and callable from the rest of the
project target.