Fix DataService
Needed a temporary provision for the meantime so that we can images in
the demo tomorrow, but that needs to happen on the server before we can
have that come from the request. Therefore, made a quick fix back to the
original functionality for now, and will add that later.