Detail quest view
Before, there was an awfully hacky way to transfer data between the
different classes in the segue towards the individual quest view, so
that was fixed. Additionally, some UI elements were added to the
application so that it starts to actually look like there's something
going on on the quest page.