Refactor Quest endpoints
So the purpose of this refactor is that the DataService
is getting a bit large, and the endpoints are starting to only really
be relevant to particular parts of the project. Due to the meeting today
with Carleton Web Services, modularizing has a clear appeal to allow for
deployment with functionality depending on a simple module format. The
DataService is one of those modules that ties these together in that
they have common dependency, it makes sense to split it up now.