Update Quest UI buttons
Originally, the buttons took up a really small portion of the screen and
were annoying to press. Decided to just make those bigger, along with
the progress bar to make it more obvious that it's present. Not sure how
to denote it so that people definitely know that's what it is for, but
maybe letting some users try it out might shed some light on that.