Cache calendar data
Currently caches by immediately spinning off a request from the
AppDelegate, and then trying to get cached data first in the
getCalendar() method inside the CalendarViewController. However, it
starts to hold off the app because I think it's using the main thread to
do that work, much like it was before, so I'm probably going to have to
figure out how to fork off a new thread to load the calendar data.