Adjust text and cells in quest view
Before, you were able to highlight cells in the quest information table
cell, which didn't make a whole lot of sense. That's been fixed. In
addition, the text now doesn't run off of the screen, but the cells
themselves don't adjust to the size of the text, which needs to be
fixed.