Fix loading speed issue
APPARENTLY using a certain type of font over another can cause HUGE
differences in the amount of time that is taken loading for the first
time for some reason. The Black version of Roboto loads much more slowly
than the Bold version. God knows why. This one was a pain to find,
that's for sure.