Memories data refactor
Realized that there were some bad practice things going on when building
responses from the server and a bunch of other things regarding memories
that could just be handled by using a specific Memory type. So I made
that change, which made a bunch of code much less painful to look at.