All geofences show up on the map at all times
Geomonitoring doesn't make as much sense for the users just wanting
to browse, so we trashed it. It's currently a bit of a hack since it
just gets rid of methods handling monitoring and doesn't get rid
of some other extraneous complexities associated with the old
layout. That will get fixed next.