Revamp tab bar, add Info tab
There were some really peculiar issues coming up with the tab
bar, so I did a revamp to make sure that it behaved consistently. Seemed
like after enough modifications some things just wouldn't work. In
addition, the info tab icon shows up only because you change it in the
view that is pointed to by the tab bar controller, and those changes
won't be reflected in the Main storyboard. That's really buggy
on Apple's part, but as long as I keep that in mind it should be good.