Add Webview, custom useragent
The plan for the info page was to provide information as needed to
different people using the app, which can be modified by Web Services as
needed. Therefore, we're simply routing to the appinfo web page, and
that can be redirected as needed. Additionally, a custom useragent was
added so that special styles can be applied for the app on each page.