Internet connection view moved out of forefront
When the internet connection overlays were set up, they were placed
directly on top of the storyboard views, and then simply shown or hidden
based on whether there was a connection or not. However, this made it
really hard to actually edit anything on the storyboard. This has been
fixed by moving it to the back and then programmatically moving it in
front of the other views when it's required.