Port to Alamofire 3.0
There were some minor breaking changes with the Alamofire library that I
had to deal with in order to switch to the latest version of Alamofire,
which was necessary to be able to use their image networking library. I
figured the asynchronous by default networking would come in handy for
when I started requesting for images from different URLs per the new
API changes.