Change deployment target to 8.2
It turned out that I was only currently supporting iOS 9.1 onward in this
project, but unfortunately that leaves out about 16% of people in
theory that haven't installed the newer OS.

I've moved back the deployment target, and realized that this involves
removing all of the storyboard references, since that was a new feature
that came with iOS 9.0 and Xcode 7. Annoying, but I figure the
deployment target can probably be moved up a year from now by whoever is
working on this next.

If you're reading this, use storyboard references if the deployment
target is iOS 9.0 or higher, since it's SO much nicer than the current
setup.