Change how data delivered to calendar
Before we were using a list with a bunch of events in it, but I now want
to be able to add headers, which will require that I know which day the
event is on. It seemed like it would be silly to sort this stuff out
after the fact, so I changed how it worked.