Change reload times
Since reunion will be the first time the app is used, we wanted the
timeouts for the data in the app to be shorter in case hotfixes needed
to be made to the data in order to make things work smoothly.

This way if a user is having issues and that is reported, changes to the
data won't take a week before the refresh automatically.