Adjust the way location services is used in the Quest
LS now check to see if it is enabled and presents appropriate error
message to the user via an alert. This handles edge cases with people
turning LS on/off after the initial request better than before.

Also, the app now only asks for permission to use LS while in use and
not “always”. This will avoid raising red flags in the app review
process